+++ src/tasks/task_id.cpp
@@ -21,6 +21,7 @@
 
 #include <algorithm>
 #include <bitset>
+#include <iostream>
 #include <stdexcept>
 #include <string>
 #include <utility>